EnergyModelsBase.jl

EnergyModelsBase — Module

Main module for EnergyModelsBase.jl.

This module provides the framework for building energy system models.

CleanExport is an operational, multi nodeal energy system model, written in Julia. The model is based on the JuMP optimization framework. It is a multi carrier energy model, where the definition of the resources are fully up to the user of the model. One of the primary design goals was to develop a model that can eaily be extended with new functionality without the need to understand and remember every variable and constraint in the model.

For running a basic energy system model, only the base technology package EnergyModelsBase.jl and the time structure package TimeStruct.jl is needed.

The main package provides simple descriptions for energy sources, sinks, conversion, and storage units. It corresponds to an operational model without geographic features.

Other packages can the optionally be added if specific functionality or technology nodes are needed. The most important packages are

EnergyModelsGeography.jl: this package makes it possible to easily extend your energy model different geographic areas, where transmission can be set to allow for the transport of resources between the different areas.

EnergyModelsInvestments.jl: this package implements functionality for investments, where the length of the investment periods are fully flexible and is decided by setting the time structure.

Packages implementing technology specific nodes:

EnergyModelsCO2.jl: implementing a CO2-storage node.

EnergyModelsHydrogen.jl: implementing an electrolyser node.

EnergyModelsRenewableProducers.jl: implements NonDisRES for intermittent renewable energy sources and RegHydroStor modeling a regulated hydro storage.

Manual outline

Creating a new node

EnergyModelsBase.jl

Manual outline

Library outline

Constraint functions

Capacity constraints

Flow constraints

Operational cost constraints

Optimization variables

General variables

Flow variables

Capacity variables

Sink variables

Other variables

Internals

Index

Types

Methods

Public interface

Resource

Nodes

Links

Model and data

Release notes

Version 0.5.1 (2023-06-16)

Version 0.5.0 (2023-06-01)

Version 0.4.0 (2023-05-30)

Version 0.3.3 (2023-04-26)

Version 0.3.2 (2023-02-07)

Version 0.3.1 (2023-02-03)

Version 0.3.0 (2023-02-02)

Version 0.2.7 (2022-12-12)

Version 0.2.4 (2022-09-07)

Version 0.2.3 (2021-09-07)

Version 0.2.2 (2021-08-20)

Version 0.2.1 (2021-04-22)

Version 0.2.0 (2021-04-19)

Version 0.1.0 (2021-03-19)

Nodes

Philosophy

General design philosophy

Description of technologies

Extensions to the model

Quick Start

Examples

The package is installed with ] add

The code was downloaded with git clone

Library outline

Public interface

Resource

Nodes

Links

Model and data

Optimization variables

General variables

Flow variables

Capacity variables

Sink variables

Other variables

Constraint functions

Capacity constraints

Flow constraints

Operational cost constraints

Internals

Index

Types

Methods