Public interface

Resource

EnergyModelsBase.Resource — Type

Resources that can be transported and converted.

Fields

id is the name/identifyer of the resource.

CO2_int::T is the the CO2 intensity.

EnergyModelsBase.ResourceCarrier — Type

General resources.

EnergyModelsBase.ResourceEmit — Type

Resources that can can be emitted (e.g., CO2, CH4, NOx).

Fields

id is the name/identifyer of the resource.

CO2_int::T is the the CO2 intensity.

Nodes

Missing docstring.

Missing docstring for Node. Check Documenter's build log for details.

EnergyModelsBase.Source — Type

Source node with only output.

EnergyModelsBase.Network — Type

Network node with both input and output.

EnergyModelsBase.Sink — Type

Sink node with only input.

EnergyModelsBase.Storage — Type

Storage node with level.

EnergyModelsBase.Availability — Type

Availability node as routing node.

Reference nodes

EnergyModelsBase.RefSource — Type

A reference Source node.

Process emissions can be included, but if the field is not added, then no process emissions are assumed through the usage of a constructor.

Fields

id is the name/identifier of the node.

Cap::TimeProfile is the installed capacity.

Opex_var::TimeProfile is the variational operational costs per energy unit produced.

Opex_fixed::TimeProfile is the fixed operational costs.

Output::Dict{Resource, Real} are the generated Resources with conversion value Real..

Data::Array{Data} is the additional data (e.g. for investments).

Emissions::Dict{ResourceEmit, Real}: emissions per energy unit produced.

EnergyModelsBase.RefNetwork — Type

A reference Network node.

Fields

id is the name/identifier of the node.

Cap::TimeProfile is the installed capacity.

Opex_var::TimeProfile is the variational operational costs per energy unit produced.

Opex_fixed::TimeProfile is the fixed operational costs.

Input::Dict{Resource, Real} are the input Resources with conversion value Real.

Output::Dict{Resource, Real} are the generated Resources with conversion value Real.

Data::Array{Data} is the additional data (e.g. for investments).

EnergyModelsBase.RefNetworkEmissions — Type

A reference Network node with process emissions.

Fields

id is the name/identifier of the node.

Cap::TimeProfile is the installed capacity.

Opex_var::TimeProfile is the variational operational costs per energy unit produced.

Opex_fixed::TimeProfile is the fixed operational costs.

Input::Dict{Resource, Real} are the input Resources with conversion value Real.

Output::Dict{Resource, Real} are the generated Resources with conversion value Real.

CO2 is required to be included the be available to have CO2 capture applied properly.

Emissions::Dict{ResourceEmit, Real}: emissions per unit produced.

CO2_capture::Real is the CO2 capture rate.

Data::Array{Data} is the additional data (e.g. for investments).

EnergyModelsBase.RefSink — Type

A reference Sink node.

This node corresponds to a demand given by the field Cap. Process emissions can be included, but if the field is not added, then no process emissions are assumed through the usage of a constructor.

Fields

id is the name/identifier of the node.

Cap::TimeProfile is the Demand.

Penalty::Dict{Any, TimeProfile} are penalties for surplus or deficits.

Requires the fields :Surplus and :Deficit.

Input::Dict{Resource, Real} are the input Resources with conversion value Real.

Emissions::Dict{ResourceEmit, Real}: emissions per unit produced.

EnergyModelsBase.RefStorage — Type

A reference Storage node.

This node is designed to store a ResourceCarrier.

Fields

id is the name/identifier of the node.

Rate_cap::TimeProfile is the installed rate capacity, that is e.g. power or mass flow.

Stor_cap::TimeProfile is the installed storage capacity, that is e.g. energy or mass.

Opex_var::TimeProfile is the variational operational costs per energy unit produced.

Opex_fixed::TimeProfile is the fixed operational costs.

Stor_res::Resource is the stored Resource.

Input::Dict{Resource, Real} are the input Resources with conversion value Real.

Output::Dict{Resource, Real} are the generated Resources with conversion value Real.

Only relevant for linking and the stored Resource.

Data::Array{Data} is the additional data (e.g. for investments).

EnergyModelsBase.RefStorageEmissions — Type

A reference Storage node.

This node is designed to store a ResourceEmit.

Fields

id is the name/identifier of the node.

Rate_cap::TimeProfile is the installed rate capacity, that is e.g. power or mass flow.

Stor_cap::TimeProfile is the installed storage capacity, that is e.g. energy or mass.

Opex_var::TimeProfile is the variational operational costs per energy unit produced.

Opex_fixed::TimeProfile is the fixed operational costs.

Stor_res::Resource is the stored Resource.

Input::Dict{Resource, Real} are the input Resources with conversion value Real.

Output::Dict{Resource, Real} are the generated Resources with conversion value Real.

Only relevant for linking and the stored Resource.

Data::Array{Data} is the additional data (e.g. for investments).

Links

EnergyModelsBase.Linear — Type

Linear Link, that is input equals output.

EnergyModelsBase.Link — Type

Declaration of the general type for links connecting nodes.

EnergyModelsBase.Direct — Type

Direct <: Link

A direct link between two nodes.

Fields

id is the name/identifier of the link.

from::Node is node from which there is flow into the link.

to::Node is node to which there is flow out of the link.

Formulation::Formulation is the used formulation of links. If not specified,

a Linear link is assumed

Model and data

EnergyModelsBase.EnergyModel — Type

Abstract type for differentation between types of models (investment, operational, ...).

EnergyModelsBase.OperationalModel — Type

Operational Energy Model without investments.

Fields

Emission_limit is a dictionary with individual emission limits as TimeProfile for each

emission resource ResourceEmit.

CO2_instance is a ResourceEmit and corresponds to the type used for CO2.

EnergyModelsBase.Data — Type

Abstract type used to define concrete struct containing the package specific elements to add to the composite type defined in this package.

EnergyModelsBase.EmptyData — Type

Empty composite type for Data