EnergyModelsGeography.jl

EnergyModelsGeography — Module

Main module for EnergyModelsGeography.jl.

Exports:

Area, RefArea, GeoAvailability, Transmission, RefStatic, RefDynamic, PipeMode, PipeSimple, PipeLinepackSimple.

This Julia package provides an operational, multi nodal model.

The package extends EnergyModelsBase.jl with functionality to set up an energy system consisting of several separate regions, with transmissions between.

Manual outline

Quick Start

Philosophy

Examples

The package is installed with ] add

The code was downloaded with git clone

Library outline

Public interface

Area

Transmission

TransmissionMode

Optimization variables

Area

TransmissionMode

Constraint functions

Capacity constraints

Transmission loss functions

Balance functions

Operational cost constraints

Reference

Index

Types

Methods