Public interface

Area

A geographical Area consist of a location and a connection to a local energy system via a specialized Availability node called GeoAvailability. Constraints related to the area keep track of a resource's export and import to the local system and exchange with other areas. Multiple dispatch is used on the Area type for imposing specific constraints. Hence, other restrictions can be applied on a area level, such as electricity generation reserves, CO<sub>2</sub> emission limits or resource limits (wind power, natural gas etc.).

EnergyModelsGeography.Area — Type

Declaration of the general type for areas.

EnergyModelsGeography.RefArea — Type

A reference Area.

Fields

id is the name/identifier of the area.

Name is the name of the area.

Lon::Real is the longitudinal position of the area.

Lat::Real is the latitudinal position of the area.

An::Availability is the Availability node routing different resources within an area.

Transmission

Transmission occurs on specified transmission corridors From one area To another. On each corridor, there can exist several TransmissionModes that are transporting resources using a range of technologies.

EnergyModelsGeography.Transmission — Type

A Transmission corridor.

A geographic corridor where TransmissionModes are used to transport resources.

Fields

From::Area is the area resources are transported from.

To::Area is the area resources are transported to.

Modes::Array{TransmissionMode} is the transmission modes that are available.

TransmissionMode

TransmissionMode describes how resources are transported, for example by dynamic transmission modes on ship, truck or railway (represented generically by RefDynamic) or by static transmission modes on overhead power lines or gas pipelines (respresented generically by RefStatic). TransmissionModes includes capacity limits (Trans_cap), losses (Trans_loss) and directions (Directions) for the generic transmission modes RefDynamic and RefStatic. More specialized TransmissionModes such as subtypes of the abstrac type PipeMode can convert one Inlet resource to another Outlet resource. The PipeMode can be Consuming another resource such as electricity for pumps at a Consumption_rate in order to transport natural gas or hydrogen.

The following TransmissionModes are implemented and exported:

EnergyModelsGeography.TransmissionMode — Type

Declaration of the general type for transmission mode transporting resources between areas.

EnergyModelsGeography.RefStatic — Type

A reference static TransmissionMode.

Generic representation of static transmission modes, such as overhead power lines or pipelines.

Fields

Name::String is the name/identifyer of the transmission mode.

Resource::Resource is the resource that is transported.

Trans_cap::Real is the capacity of the transmission mode.

Trans_loss::Real is the loss of the transported resource during transmission, modelled as a ratio.

Opex_var::TimeProfile is the variational operational costs per energy unit transported.

Opex_fixed::TimeProfile is the fixed operational costs per installed capacity.

Directions is the number of directions the resource can be transported, 1 is unidirectional (A->B) or 2 id bidirectional (A<->B).

Data::Dict{String, Data} is the additional data (e.g. for investments).

EnergyModelsGeography.RefDynamic — Type

A reference dynamic TransmissionMode.

Generic representation of dynamic transmission modes, using for example truck, ship or railway transport.

Fields

Name::String is the name/identifyer of the transmission mode.

Resource::Resource is the resource that is transported.

Trans_cap::TimeProfile is the capacity of the transmission mode.

Trans_loss::TimeProfile is the loss of the transported resource during transmission, modelled as a ratio.

Opex_var::TimeProfile is the variational operational costs per energy unit transported.

Opex_fixed::TimeProfile is the fixed operational costs per installed capacity.

Directions is the number of directions the resource can be transported, 1 is unidirectional (A->B) or 2 id bidirectional (A<->B).

Data::Dict{String, Data} is the additional data (e.g. for investments).

EnergyModelsGeography.PipeMode — Type

TransmissionMode mode for additional variable potential.

EnergyModelsGeography.PipeSimple — Type

This TransmissionMode allows for altering the transported Resource.

A usage of this could e.g. be by defining a subtype struct of Resource with the field 'pressure'. This PipelineMode can then take SomeSubtype<:Resource with pressure p₁ at the inlet, and pressure p₂ at the outlet.

This type also supports consuming resources proportionally to the volume of transported Resource (at the inlet). This could be used for modeling the power needed for operating the pipeline.

Fields

Name::String is the identifier used in printed output.

Inlet::Resource is the Resource going into transmission.

Outlet::Resource is the Resource going out of the outlet of the transmission.

Consuming::Resource is the Resource the transmission consumes by operating.

Consumption_rate::Real the rate of which the resource Pipeline.Consuming is consumed, as a ratio of the volume of the resource going into the inlet. I.e.:

  `Consumption_rate` = consumed volume / inlet volume (per operational period)

Trans_cap::Real is the capacity of the transmission mode.

Trans_loss::Real is the loss of the transported resource during transmission, modelled as a ratio.

Opex_var::TimeProfile is the variational operational costs per energy unit transported.

Opex_fixed::TimeProfile is the fixed operational costs per installed capacity.

Directions specifies that the pipeline is Unidirectional (1) by default.

Data::Dict{String, Data} is the additional data (e.g. for investments).

EnergyModelsGeography.PipeLinepackSimple — Type

PipeLinepackSimple <: TransmissionMode

Pipeline model with linepacking implemented as simple storage function.

Fields (additional to PipeSimple)

Linepack_energy_share::Flaot64 - is the storage energy capacity relative to pipeline capacity.

Level_share_init::Float64 - is the initial storage level.

Data::Dict{String, Data} is the additional data (e.g. for investments).