Public interface

InvestmentModel

This structure defines a new type of Energy Model than the one available from EnergyModelsBase. This new structure is used when investment needs to be part of the analysis. This struct is a subtype of AbstractInvestmentModel which is itself a subtype of EMB.EnergyModel.

The structure contains some key information for the model such as the emissions limits and penalties for each ResourceEmit. The discout rate is an important element of investment analysis needed to represent the present value of future cash flows. It is provided to the model as a value between 0 and 1 ( e.g. a discount rate of 5% is 0.05).

Investment Types

Different investment types are available to help represent different situations. The investment type is defined in Inv_mode in extra_inv_data and extra_inv_data_storage. The other fields in those struct are used to define the relevant parameters of the different investment types.

BinaryInvestment

BinaryInvestment is a type of investment that can either be installed, or not, at the defined capacity. The capacity of the investment cannot be adjusted by the optimization.

Note

This investment type leads to the addition of binary variables.

DiscreteInvestment

DiscreteInvestment is a type of investment where a technology can be invested in increments of a given size. It can for example be used to represent investment in modular technologies that can be scaled by adding several modules together. The field Cap_increment::TimeProfile from extra_inv_data, or Rate_increment::TimeProfile and Stor_increment::TimeProfile from extra_inv_data_storage should be provided.

Note

This investment type leads to the addition of integer variables.

ContinuousInvestment

ContinuousInvestment is a type of investment where the investment in the capacity happens linearly from 0 and up to a given Cap_max_inst::TimeProfile. The rate of installation of the technology can also be limited with Cap_max_add::TimeProfile

Note

Defining Cap_min_add::TimeProfile for this type of investment will lead to a forced investment of at least Cap_min_add in each period.

SemiContinuousInvestment

SemiContinuousInvestment is a type of investment similar to ContinuousInvestment, but where the investment is either 0 or between a minimum and maximum value. This means you can define Cap_min_add::TimeProfile without forcing investment in the technology.

Note

This investment type leads to the addition of binary variables.

SemiContinuousOffsetInvestment

SemiContinuousOffsetInvestment is a type of investment similar to SemiContinuousInvestment and implemented for investments in transmission infrastructure. It does differ with respect to how the costs are calculated. A SemiContinuousOffsetInvestment has an offset in the cost implemented through the the field Capex_trans_offset. This offset corresponds to the theoretical cost at an invested capacity of 0.

Note

This investment type leads to the addition of binary variables.

FixedInvestment

FixedInvestment is a type of investment where an investment in the given capacity is forced. It allows to account for the investment cost of known investments.

LifetimeMode

Several ways to define the lifetime of a technology are available in the package and presented below.

UnlimitedLife

This LifetimeMode is used when the lifetime of an asset is not limited. No reinvestment is considered by the optimization and there is also ne salvage value (or rest value) in the last period.

StudyLife

This LifetimeMode is used to define that the investment should be available for the whole study period. That means that this technology will be reinvested in at the end of its lifetime, and as many times as necessary to reach the end of the study. A rest value is also calculated for the remaining years at the end of the study period.

PeriodLife

This LifetimeMode is used to define that the investment is only lasting for the strategic period in which it happens. Additional year of lifetime are counted as a rest value. Reinvestment inside the strategic periods are also considered in case the lifetime is shorter than the length of the startegic period.

RollingLife

This LifetimeMode is used to define that the investment rolls over to the next strategic periods as long as the lifetime is not reached. If the remaining lifetime falls between two strategic periods, the investment is not carried to the next period and a rest value is used t oaccount for the remaining lifetime.

Additional Data for Investments

Additional data for investment is specified when creating the nodes. Two struct are used to define the paramters necessary for production technologies (InvData) and storages (InvDataStorage)

InvData

Define the structure for the additional parameters passed to the technology structures defined in other packages. It uses Base.@kwdef to use keyword arguments and default values. The name of the parameters have to be specified.

!!!note Depending on the type of investment mode chosen, not all parameters are necessary. It is however possible to set parameters even if they will not be used t obe able to change the investment type more easily.

InvDataStorage

Define the structure for the additional parameters passed to the technology structures defined in other packages. It uses Base.@kwdef to use keyword arguments and default values. The name of the parameters have to be specified. The parameters are separated between Rate and Stor. The Rate refers to instantaneous component (Power, Flow, ...) for instance, charging and discharging power for batteries, while the Stor refers to a volumetric component (Energy, Volume, ...), for instance storage capacity for a battery.

!!!note Depending on the type of investment mode chosen, not all parameters are necessary. It is however possible to set parameters even if they will not be used t obe able to change the investment type more easily.

TransInvData

Similarly as for InvData, this struct defines additional parameters necessary for handling the investment in transmission between geographical areas. This struct is used in addition to EnergyModelsGeography to add investment in transmission.

Index

EnergyModelsInvestments.BinaryInvestment

EnergyModelsInvestments.ContinuousInvestment

EnergyModelsInvestments.DiscreteInvestment

EnergyModelsInvestments.FixedInvestment

EnergyModelsInvestments.InvData

EnergyModelsInvestments.InvDataStorage

EnergyModelsInvestments.InvestmentData

EnergyModelsInvestments.InvestmentModel

EnergyModelsInvestments.LifetimeMode

EnergyModelsInvestments.PeriodLife

EnergyModelsInvestments.RollingLife

EnergyModelsInvestments.SemiContinuousInvestment

EnergyModelsInvestments.SemiContinuousOffsetInvestment

EnergyModelsInvestments.StudyLife

EnergyModelsInvestments.TransInvData

EnergyModelsInvestments.UnlimitedLife

Types

EnergyModelsInvestments.BinaryInvestment — Type

Binary investment in given capacity with binary variables. Requires specification of Cap_start in InvData for proper analyses.

EnergyModelsInvestments.ContinuousInvestment — Type

Continuous investment between zero and a maximum value.

EnergyModelsInvestments.DiscreteInvestment — Type

Investment in fixed increments with integer variables.

EnergyModelsInvestments.FixedInvestment — Type

Forced investment in given capacity.

EnergyModelsInvestments.InvData — Type

Extra data for investing in technologies.

Define the structure for the additional parameters passed to the technology structures defined in other packages. It uses Base.@kwdef to use keyword arguments and default values. The name of the parameters have to be specified.

Fields

Capex_Cap::TimeProfile Capital Expenditure for the capacity, here investment costs of

the technology in each period.

Cap_max_inst::TimeProfile Maximum possible installed capacity of the technology in

each period.

Cap_max_add::TimeProfile Maximum capacity addition in one period from the previous.

Cap_min_add::TimeProfile Minimum capacity addition in one period from the previous.

Inv_mode::Investment = ContinuousInvestment() Type of the investment:

BinaryInvestment, DiscreteInvestment, ContinuousInvestment, SemiContinuousInvestment, or FixedInvestment.

Cap_start::Union{Real, Nothing} = nothing Starting capacity in first period.

If nothing is given, it is set by get_start_cap() to the capacity Cap of the node.

Cap_increment::TimeProfile = FixedProfile(0) Capacity increment used in the case of

DiscreteInvestment

Life_mode::LifetimeMode = UnlimitedLife() Type of handling of the lifetime:

UnlimitedLife, StudyLife, PeriodLife or RollingLife

Lifetime::TimeProfile = FixedProfile(0) Duration/Lifetime of the technology invested

in each period.

EnergyModelsInvestments.InvDataStorage — Type

Extra data for investing in storages.

Define the structure for the additional parameters passed to the technology structures defined in other packages. It uses Base.@kwdef to use keyword arguments and default values. The name of the parameters have to be specified. The parameters are separated between Rate and Stor. The Rate refers to instantaneous component (Power, Flow, ...) for instance, charging and discharging power for batteries, while the Stor refers to a volumetric component (Energy, Volume, ...), for instance storage capacity for a battery.

Fields

Capex_rate::TimeProfile Capital Expenditure for storage rate, here investment

costs of the technology rate in each period.

Rate_max_inst::TimeProfile Maximum possible installed rate of the technology in

each period.

Rate_max_add::TimeProfile Maximum rate addition in one period from the previous.

Rate_min_add::TimeProfile Minimum rate addition in one period from the previous.

Capex_stor::TimeProfile Capital Expenditure, here investment costs of the technology

storage volume in each period.

Stor_max_inst::TimeProfile Maximum possible installed storage volume of the technology

in each period.

Stor_max_add::TimeProfile Maximum storage volume addition in one period from the

previous.

Stor_min_add::TimeProfile Minimum storage volume addition in one period from the

previous.

Inv_mode::Investment = ContinuousInvestment() Type of the investment:

BinaryInvestment, DiscreteInvestment, ContinuousInvestment, SemiContinuousInvestment or FixedInvestment.

Rate_start::Union{Real, Nothing} = nothing Starting rate in first period.

If nothing is given, it is set by get_start_cap() to the capacity Rate_cap of the node.

Stor_start::Union{Real, Nothing} = nothing Starting storage volume in first period.

If nothing is given, it is set by get_start_cap() to the capacity Stor_cap of the node.

Rate_increment::TimeProfile = FixedProfile(0) Rate increment used in the case of

DiscreteInvestment

Stor_increment::TimeProfile = FixedProfile(0) Storage volume increment used in the

case of DiscreteInvestment

Life_mode::LifetimeMode = UnlimitedLife() Type of handling of the lifetime:

UnlimitedLife, StudyLife, PeriodLife, or RollingLife

Lifetime::TimeProfile = FixedProfile(0) Duration/Lifetime of the technology invested

in each period.

EnergyModelsInvestments.InvestmentData — Type

Abstract type for the extra data for investing in technologies.

EnergyModelsInvestments.InvestmentModel — Type

An concrete basic investment model type

Fields

Emission_limit::Dict{ResourceEmit, TimeProfile} are the emission caps for the different

emissions types considered.

Emission_price::Dict{ResourceEmit, TimeProfile} are the prices for the different

emissions types considered.

CO2_instance is a ResourceEmit and corresponds to the type used for CO2.

r is the discount rate in the investment optimization.

EnergyModelsInvestments.LifetimeMode — Type

Abstract lifetime type

EnergyModelsInvestments.PeriodLife — Type

The investment is considered to last only for the strategic period. The excess lifetime is considered in the rest value. If the lifetime is lower than the length of the period, reinvestment is considered as well.

EnergyModelsInvestments.RollingLife — Type

The investment is rolling to the next strategic periods and it is retired at the end of its lifetime or the end of the previous strategic period if its lifetime ends between two periods.

EnergyModelsInvestments.SemiContinuousInvestment — Type

Semi-continuous investment where the cost is going through the origin.

EnergyModelsInvestments.SemiContinuousOffsetInvestment — Type

Semi-continuous investment where the cost has an additional offset

EnergyModelsInvestments.StudyLife — Type

The investment lasts for the whole study period with adequate reinvestments at the end of the lifetime and considering the rest value.

EnergyModelsInvestments.TransInvData — Type

Extra data for investing in transmission.

Define the structure for the additional parameters passed to the technology structures defined in other packages It uses Base.@kwdef to use keyword arguments and default values. The name of the parameters have to be specified.

Fields

Capex_trans::TimeProfile Capital Expenditure for the transmission capacity, here investment costs of the transmission in each period.

Trans_max_inst::TimeProfile Maximum possible installed transmission capacity in each period.

Trans_max_add::TimeProfile Maximum transmission capacity addition in one period from the previous.

Trans_min_add::TimeProfile Minimum transmission capacity addition in one period from the previous.

Inv_mode::Investment = ContinuousInvestment() Type of the investment: BinaryInvestment, DiscreteInvestment, ContinuousInvestment, SemiContinuousInvestment or FixedInvestment.

Trans_start::Union{Real, Nothing} = nothing Starting transmission capacity in first period. If nothing is given, it is set by getstartcap() to the capacity Trans_Cap of the transmission.

Trans_increment::TimeProfile = FixedProfile(0) Transmission capacity increment used in the case of DiscreteInvestment

EnergyModelsInvestments.UnlimitedLife — Type

The investment's life is not limited. The investment costs do not consider any reinvestment or rest value.

Methods